third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/settings/ignorelist-a11y-test-expected.txt

Tests accessibility in IgnoreList view using the axe-core linter.

Running: testAddPattern
Opened input box: true
aXe violations: []


Running: testPatternList
Added a pattern in the list: (Default),test*
aXe violations: []


Running: testPatternError
Error message: Rule can't be empty
aXe violations: []
